requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
requests==2.22.0
Wand==0.6.3
redis==3.4.1
rq==1.2.2
boto3==1.16.20
redis==3.4.1
pdf2image==1.14.0
pdf2pptx==1.0.3
awscli==1.19.98
APScheduler==3.7.0
pdf2docx==0.5.0
python-dotenv==0.18.0
img2pdf==0.4.1
Flask-Babel==2.0.0
Flask-Minify==0.32